tests: fix
