+++ test/config-format/commonjs-default/main.js
@@ -1,0 +1,1 @@
+console.log('Hoshiumi');
