+++ test/config-format/commonjs-default/webpack.config.cjs
@@ -1,0 +1,12 @@
+const path = require('path');
+
+const config = {
+    mode: 'production',
+    entry: './main.js',
+    output: {
+        path: path.resolve(__dirname, 'dist'),
+        filename: 'foo.bundle.js',
+    },
+};
+
+module.exports.default = config;
