--- test/config-format/typescript-commonjs-default/main.ts
@@ -1,1 +1,0 @@
-console.log('Main typescript file');
