--- test/config-format/typescript-commonjs-default/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs"
-  }
-}
