--- test/config-format/typescript-commonjs-default/typescript-commonjs-default.test.js
@@ -1,14 +1,0 @@
-const { run } = require('../../utils/test-utils');
-const { existsSync } = require('fs');
-const { resolve } = require('path');
-
-describe('webpack cli', () => {
-    it('should support typescript file', () => {
-        const { exitCode, stderr, stdout } = run(__dirname, ['-c', './webpack.config.ts']);
-
-        expect(stderr).toBeFalsy();
-        expect(stdout).toBeTruthy();
-        expect(exitCode).toBe(0);
-        expect(existsSync(resolve(__dirname, 'dist/foo.bundle.js'))).toBeTruthy();
-    });
-});
