+++ test/config-format/typescript/main.ts
