+++ test/config-format/typescript/tsconfig.json
