+++ test/config-format/typescript/typescript-commonjs.test.js
